refactor: 画像を更新する処理時のクエリを改善

+++ app/forms/article_form.rb
@@ -25,14 +25,16 @@ class ArticleForm
     ActiveRecord::Base.transaction do
       article.update!(title: title, body: body)
 
-      if destroying_image_ids.present?
-        destroying_images = article.images.where(id: destroying_image_ids)
-        destroying_images.delete_all
-        article.reload
-      end
+      destroying_images = article.images.where(id: destroying_image_ids)
+      destroying_images.delete_all if destroying_images.present?
+      article.reload
 
-      updating_images.each { |image| article.images.find(image['id']).update!(position: image['position']) }
-      new_images.each { |image| article.images.create!(cl_id: image['cl_id'], position: image['position']) }
+      updating_images = article.images.where(id: updating_image_ids).reorder(:id)
+      updating_images.zip(updating_image_positions) { |image, position| image.update!(position: position) }
+
+      new_image_attributes_collection.each do |attrs|
+        article.images.create!(cl_id: attrs['cl_id'], position: attrs['position'])
+      end
     end
 
     true
@@ -45,22 +47,34 @@ class ArticleForm
     article
   end
 
-  def updating_images
+  def updating_image_attributes_collection
     return [] if image_attributes.nil?
 
-    image_attributes.values.filter { |v| v['id'] && v['_destroy'] == 'false' }
+    image_attributes.values.filter { |attrs| attrs['_destroy'] == 'false' }
   end
 
-  def new_images
+  def updating_image_ids
     return [] if image_attributes.nil?
 
-    image_attributes.values.filter { |v| v['cl_id'] }
+    updating_image_attributes_collection.map { |attrs| attrs['id'] }
+  end
+
+  def updating_image_positions
+    return [] if image_attributes.nil?
+
+    updating_image_attributes_collection.sort_by { |attrs| attrs['id'] }.map { |attrs| attrs['position'] }
+  end
+
+  def new_image_attributes_collection
+    return [] if image_attributes.nil?
+
+    image_attributes.values.filter { |attrs| attrs['cl_id'] }
   end
 
   def destroying_image_ids
     return [] if image_attributes.nil?
 
-    image_attributes.values.filter { |v| v['_destroy'] == 'true' }.map { |v| v['id'] }
+    image_attributes.values.filter { |attrs| attrs['_destroy'] == 'true' }.map { |attrs| attrs['id'] }
   end
 
   def saved_images
